feat: add SignUp service
