+++ .idea/misc.xml
@@ -3,20 +3,13 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
-        <entry key="..\:/Studentmarket/StudentMarket/app/src/main/res/drawable/accounticon.xml" value="0.1435" />
-        <entry key="..\:/Studentmarket/StudentMarket/app/src/main/res/drawable/button_color.xml" value="0.174" />
-        <entry key="..\:/Studentmarket/StudentMarket/app/src/main/res/drawable/categoryicon.xml" value="0.1435" />
         <entry key="..\:/Studentmarket/StudentMarket/app/src/main/res/drawable/edittextstyle.xml" value="0.25" />
         <entry key="..\:/Studentmarket/StudentMarket/app/src/main/res/drawable/gradient_background.xml" value="0.25" />
-        <entry key="..\:/Studentmarket/StudentMarket/app/src/main/res/drawable/homeIcon.xml" value="0.129" />
-        <entry key="..\:/Studentmarket/StudentMarket/app/src/main/res/drawable/homeicon.xml" value="0.1215" />
         <entry key="..\:/Studentmarket/StudentMarket/app/src/main/res/drawable/ic_baseline_account_circle_24.xml" value="0.25" />
         <entry key="..\:/Studentmarket/StudentMarket/app/src/main/res/drawable/ic_baseline_category_24.xml" value="0.25" />
         <entry key="..\:/Studentmarket/StudentMarket/app/src/main/res/drawable/ic_baseline_favorite_24.xml" value="0.25" />
         <entry key="..\:/Studentmarket/StudentMarket/app/src/main/res/drawable/ic_baseline_home_24.xml" value="0.25" />
-        <entry key="..\:/Studentmarket/StudentMarket/app/src/main/res/drawable/ic_close.xml" value="0.129" />
         <entry key="..\:/Studentmarket/StudentMarket/app/src/main/res/drawable/mybutton.xml" value="0.25" />
-        <entry key="..\:/Studentmarket/StudentMarket/app/src/main/res/drawable/nav_item_color_state.xml" value="0.129" />
         <entry key="..\:/Studentmarket/StudentMarket/app/src/main/res/drawable/rounded.xml" value="0.25" />
         <entry key="..\:/Studentmarket/StudentMarket/app/src/main/res/layout/activity_main.xml" value="0.264" />
         <entry key="..\:/Studentmarket/StudentMarket/app/src/main/res/layout/fragment_account.xml" value="0.2" />
@@ -28,13 +21,6 @@
         <entry key="..\:/Studentmarket/StudentMarket/app/src/main/res/layout/fragment_login2.xml" value="0.25" />
         <entry key="..\:/Studentmarket/StudentMarket/app/src/main/res/layout/fragment_pre_register.xml" value="0.2" />
         <entry key="..\:/Studentmarket/StudentMarket/app/src/main/res/layout/fragment_register.xml" value="0.25" />
-        <entry key="..\:/Studentmarket/StudentMarket/app/src/main/res/layouts/Account/layout/fragment_account.xml" value="0.25" />
-        <entry key="..\:/Studentmarket/StudentMarket/app/src/main/res/layouts/Account/layout/fragment_change_pass.xml" value="0.25" />
-        <entry key="..\:/Studentmarket/StudentMarket/app/src/main/res/layouts/Account/layout/fragment_change_password.xml" value="0.1537037037037037" />
-        <entry key="..\:/Studentmarket/StudentMarket/app/src/main/res/layouts/Account/layout/fragment_forgot_password.xml" value="0.25" />
-        <entry key="..\:/Studentmarket/StudentMarket/app/src/main/res/layouts/Account/layout/fragment_login.xml" value="0.22777777777777777" />
-        <entry key="..\:/Studentmarket/StudentMarket/app/src/main/res/layouts/Account/layout/fragment_pre_register.xml" value="0.22777777777777777" />
-        <entry key="..\:/Studentmarket/StudentMarket/app/src/main/res/layouts/Account/layout/fragment_register.xml" value="0.22777777777777777" />
         <entry key="..\:/Studentmarket/StudentMarket/app/src/main/res/menu/bottom_menu.xml" value="0.23703703703703705" />
         <entry key="..\:/Users/Dell/Documents/GitHub/StudentMarket/app/src/main/res/layout/activity_main.xml" value="0.10885416666666667" />
         <entry key="..\:/Users/Dell/Documents/GitHub/StudentMarket/app/src/main/res/layout/fragment_abc.xml" value="0.15052083333333333" />
@@ -49,6 +35,9 @@
         <entry key="..\:/Users/Dell/Documents/GitHub/StudentMarket/app/src/main/res/menu/bottom_menu.xml" value="0.12685185185185185" />
         <entry key="app/src/main/res/layouts/Account/layout/fragment_account.xml" value="0.225" />
         <entry key="app/src/main/res/layouts/Account/layout/fragment_change_pass.xml" value="0.225" />
+        <entry key="app/src/main/res/layouts/Account/layout/fragment_change_password.xml" value="0.15208333333333332" />
+        <entry key="app/src/main/res/layouts/Account/layout/fragment_forgot_password.xml" value="0.19947916666666668" />
+        <entry key="app/src/main/res/layouts/Account/layout/fragment_login.xml" value="0.19947916666666668" />
         <entry key="app/src/main/res/layouts/Account/layout/fragment_pre_register.xml" value="0.225" />
         <entry key="app/src/main/res/layouts/Account/layout/fragment_register.xml" value="0.18" />
         <entry key="app/src/main/res/xml/network_security_config.xml" value="0.1" />
