+++ app/src/main/java/com/example/studentmarket/Helper/Responsive/Responsive.java
@@ -1,0 +1,4 @@
+package com.example.studentmarket.Helper.Responsive;
+
+public class Responsive {
+}
