+++ app/src/main/java/com/example/studentmarket/Helper/Validation/Validate.java
@@ -1,4 +1,4 @@
-package com.example.studentmarket;
+package com.example.studentmarket.Helper.Validation;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
