--- app/src/main/java/com/example/studentmarket/Service/ProductService.java
@@ -1,4 +1,0 @@
-package com.example.studentmarket.Service;
-
-public class ProductService {
-}
