--- app/src/main/java/com/example/studentmarket/Service/ProfileService.java
@@ -1,5 +1,0 @@
-package com.example.studentmarket.Service;
-
-public class ProfileService {
-
-}
