--- app/src/main/java/com/example/studentmarket/Service/UserActionService.java
@@ -1,4 +1,0 @@
-package com.example.studentmarket.Service;
-
-public class UserActionService {
-}
