+++ app/src/main/java/com/example/studentmarket/Services/ProductService.java
@@ -1,0 +1,4 @@
+package com.example.studentmarket.Services;
+
+public class ProductService {
+}
