+++ app/src/main/java/com/example/studentmarket/Services/ProfileService.java
@@ -1,0 +1,5 @@
+package com.example.studentmarket.Services;
+
+public class ProfileService {
+
+}
