+++ app/src/main/java/com/example/studentmarket/Services/UserActionService.java
@@ -1,0 +1,4 @@
+package com.example.studentmarket.Services;
+
+public class UserActionService {
+}
