+++ app/src/main/res/layouts/Account/layout/fragment_pre_register.xml
@@ -68,7 +68,7 @@
                 android:layout_height="wrap_content"
                 android:layout_marginTop="42dp"
                 android:enabled="false"
-                android:text="Tiêp tục"
+                android:text="@string/common_continue"
                 app:layout_constraintTop_toBottomOf="@+id/register_password_edit_text" />
 
             <LinearLayout
