+++ app/src/main/res/values/strings.xml
@@ -17,4 +17,6 @@
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
     <string name="userFullName">Họ và tên</string>
+    <string name="SignUp_Error">SignUp fail, Please try again</string>
+    <string name="common_continue">Tiếp tục</string>
 </resources>
