+++ app/src/main/res/xml/network_security_config.xml
@@ -1,9 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <network-security-config>
-    <domain-config>
-        <domain includeSubdomains="true">http://10.0.2.2/user/signUp</domain>
-        <trust-anchors>
-
-        </trust-anchors>
+    <base-config cleartextTrafficPermitted="true" />
+    <domain-config cleartextTrafficPermitted="true">
+        <domain includeSubdomains="true">http://10.0.2.2:8080</domain>
     </domain-config>
 </network-security-config>
